+++ bliss/estimators/spectral_kurtosis.cpp
@@ -6,7 +6,7 @@
 
 using namespace bliss;
 
-bland::ndarray bliss::estimate_spectral_kurtosis(const bland::ndarray &spectrum_grid, int64_t N, int64_t M, float d) {
+bland::ndarray bliss::estimate_spectral_kurtosis(const bland::ndarray &spectrum_grid, int32_t N, int32_t M, float d) {
     fmt::print("INFO: spec kurtosis with M={} and N={}\n", M, N);
     auto s1 = bland::square(bland::sum(spectrum_grid, {0}));
     auto s2 = bland::sum(bland::square(spectrum_grid), {0});
